+++ R/resample.R
@@ -1546,7 +1546,7 @@ ResampleBioticAssignmentByStratum <- function(BioticAssignment, Seed) {
 #' 
 #' @export
 #' 
-ResampleBioticAssignmentByPSU <- function(BioticAssignment, Seed) {
+ResampleBioticAssignmentByAcousticPSU <- function(BioticAssignment, Seed) {
     
     # This function will be renamed to ResampleAssignedHaulsInStratum
     
@@ -2128,7 +2128,7 @@ initialAggregateBootstrapNetCDF4DataOne <- function(
     return(output)
 }
 
-    
+
 
 
 #' List files in a folder
@@ -2138,8 +2138,19 @@ initialAggregateBootstrapNetCDF4DataOne <- function(
 #' @export
 #' 
 ListFilesInFolder <- function(
-        FolderName = character()
+    FolderName = character(), 
+    FileExtension = character()
 ) {
-    data.table::data.table(Files = list.files(FolderName))
+    files <- list.files(FolderName)
+    ext <- tools::file_ext(files)
+    if(length(FileExtension)) {
+        files <- files[ext == FileExtension]
+    }
+    
+    return(data.table::data.table(FileNames = files))
 }
+
+
+
+