+++ NEWS.md
@@ -10,6 +10,7 @@
 * Moved definitions of StoxDateTimeFormat, emptyStratumPolygon and emptyStratumPolygonGeojson to RstoxData and RstoxBase.
 * Added the arguments check.columnNames_identical and testAllTRUE to compareProjectToStoredOutputFiles().
 * Temporarily added ListFilesInFolder() to test selecting a folder in the GUI.
+* Renamed ResampleBioticAssignmentByPSU to ResampleBioticAssignmentByAcousticPSU
 
 
 # RstoxFramework v3.6.3-9001 (2023-08-31)